Updated files -Nick
Updated a few files like the edit page so it can have the same nav bar as the rest of the admin page, also tried making a login function but no success. I changed up the services.php file as well, gave it a proper nav bar so you can navigate to the other pages and I added another bit of validation to the contact us form so that the user as to input .com or whatever they use.